+++ fedbiomed/common/metrics.py
@@ -1,5 +1,5 @@
 """
-???
+Provide test metrics, both MetricTypes to use in TrainingArgs but also calculation routines.
 """